Rename to panel.ts
